JAVA -- Linked List User Address Book
The basic idea of this project was to introduce creation of linked lists
and interfaces. This project was built using previous knowledge of JAVA
such as the MVC model, reading/writing/manipulating text files, parsing
data, and sorting data.
The basics of JAVA should already be understood when reading this code.

Program Synopsis:
A user modified address book which can add/delete/modify/display the
contents. The user is prompted with a GUI with multiple selections. When
the address book is modified with a person's entry (name, address, date,
ect.), the contents are saved to a text file which can be displayed in
order.

MVC - model, view, controller idea was also introduced.

Model classes - Maintains the state and data that the application
represents - holds the core of the application. If changes are made in
the model, it updates all of its views.

Controller class - This typically holds the use cases/main method.

View classes - The user interface which displays information about the
model to the user.